ESP32Joystick - Update commit - Hour 4
Decided to revisit this project. Since esp32-s2 doesn't really have good documentations on joystick support, i decided to use an Arduino pro micro and a esp8266 for wireless connectivity. In this code I updated the html, and condensed it, and created a basic webserver on esp8266 with the condensed html.